Ex_02_02_18 -> Ex_05_03_06 (+ WIP Ex_05_03_17)

+++ C_Fundamentals_Exercises/Ex_05_02_17/Ex_05_02_17.c
@@ -32,7 +32,7 @@ int main(void)
 	}
 	else // March till december
 	{
-		factor = 365 * year + day + 31 * (month - 1) - (0.4 * month + 2.3) + (year / (float)4) - (((year / (float)100) + 1) * (3 / 4));
+		factor = 365 * year + day + 31 * (month - 1) - (0.4 * month + 2.3) + (year / (float)4) - (int)((int)((year / (float)100) + 1) * (3 / 4));
 	}
 
 	weekDay = factor - (int)(factor / 7) * 7;
